+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ui
 
+## 1.50.1-next.1
+
+### Patch Changes
+
+- Updated dependencies [cfa4185]
+  - @copilotkit/react-core@1.50.1-next.1
+  - @copilotkit/runtime-client-gql@1.50.1-next.1
+  - @copilotkit/shared@1.50.1-next.1
+
 ## 1.50.1-next.0
 
 ### Patch Changes